requirements-dev.txt
-r requirements.txt
django-debug-toolbar
factory-boy
pytest-django